Implement support of cpuid instruction + include folder + some updates

+++ include/types.h
@@ -2,6 +2,6 @@
 
 
 typedef unsigned char uint8;
-typedef unsigned int uint16;
+typedef unsigned short int uint16;
 typedef unsigned long uint32;
 typedef unsigned long long uint64;
